+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - CHANGED: explicitly specifying encryption keys prevents encrypting for other recipients
   - GPG encryption for arbitrary keys only possible #9 #14 
 - CHANGED: Optional parameter "email_or_list" renamed to "email_or_more" (methods `to` and friends)
-  - When setting recipients, apart from list other iterables (tuple, generator, set and frozenset) may be used
+  - When setting recipients, apart from list other iterables (tuple, generator, set, ...) may be used
 - fix: when signing key fails, do not sign with the default GPG private key
 
 ## 1.5.4 (2022-04-04)